requirements.txt
absl-py==0.9.0
natsort
amqp==2.4.2
aocr==0.7.6
asn1crypto==0.24.0
astor==0.8.1
autopep8==1.3.4
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
billiard==3.5.0.5
bleach==1.5.0
ccsm==0.9.13.1
celery==4.2.1
chardet==3.0.4
cliapp==1.20170827
cloudpickle==1.3.0
cmdtest==0.32
compizconfig-python==0.9.13.1
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
Distance==0.1.3
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
gast==0.3.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-pasta==0.2.0
googleapis-common-protos==1.51.0
google-pasta==0.2.0
grpcio==1.27.2
gyp==0.1
h5py==2.10.0
html5lib==0.9999999
idna==2.6
ipaddress==1.0.17
ipython==5.5.0
ipython-genutils==0.2.0
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyboard==0.13.3
keyring==10.6.0
keyrings.alt==3.0
kiwi==1.9.22
kiwisolver==1.1.0
kombu==4.4.0
Markdown==3.1.1
matplotlib==2.2.5
mock==3.0.5
networkx==2.2
numpy==1.16.6
oauth2client==4.1.3
olefile==0.46
opencv-python==4.2.0.32
pandas==0.24.2
parsel==1.5.2
pathlib2==2.3.0
pbr==5.4.4
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.2
prompt-toolkit==1.0.15
protobuf==3.0.0b2
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pygame==1.9.6
Pygments==2.2.0
pygobject==3.26.1
pygraphviz==1.4rc1
pyparsing==2.4.6
pytesseract
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
scandir==1.7
scikit-image==0.14.5
scipy==1.2.3
SecretStorage==2.3.1
selenium==3.141.0
simplegeneric==0.8.1
six==1.14.0
stevedore==1.31.0
subprocess32==3.5.4
tensorboard==1.14.0
tensorflow==0.8.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
tqdm==4.43.0
traitlets==4.3.2
ttystatus==0.38
unity-lens-photos==1.0
vine==1.2.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.7
Werkzeug==1.0.0
wrapt==1.12.1
